+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Updated pinejs-client to ^4.0.0
 * Converted src/server/src/server-glue/server.coffee to typescript.
 * Converted src/server/src/passport-pinejs/passport-pinejs.coffee to typescript.
 * Converted src/server/src/server-glue/module.coffee to typescript.