+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Only run rules if a field they query has been attempted to be changed.
 * Updated pinejs-client to ^4.0.0
 * Converted src/server/src/server-glue/server.coffee to typescript.
 * Converted src/server/src/passport-pinejs/passport-pinejs.coffee to typescript.